Add the remaining theming
