+++ src/index.ts
@@ -10,12 +10,13 @@ export {
   ErrorProvider,
   useError,
 } from "./utils";
-export type { Theme } from "./utils/themeProvider";
+
 export type {
   ExtraEditorProps,
   Layout,
   Entry,
   Options,
+  Theme,
   ExplorerEntry,
   EditorsEntry,
   ViewersEntry,
