+++ src/types/index.ts
@@ -14,3 +14,4 @@ export type { EventData, AccentColorEventData } from "./events";
 export type { Error } from "./error";
 export { ErrorType } from "./error";
 export type { Options } from "./options";
+export type { Theme } from "./theme";
